backend/requirements.txt
fastapi==0.128.6
uvicorn[standard]==0.34.0
gunicorn==21.2.0  # 新增
google-genai==0.8.0
requests==2.31.0
# pydantic version is managed by fastapi
python-multipart==0.0.6
python-dotenv>=0.19.0
python-jose[cryptography]>=3.3.0
websockets==14.2

# 数据库依赖
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# 邮件服务依赖
resend>=1.0.0

# 测试依赖
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# 代码质量工具
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.11.2
ruff==0.9.4
pre-commit==3.7.1